requirements.txt
click==8.1.3
Flask==2.0.0
Flask_SQLAlchemy==2.5.1
PyMySQL==1.0.2
pyperclip==1.8.2
python-dotenv==0.21.0
setuptools==63.1.0
pre-commit
